Restructure into crate
