+++ Cargo.toml
@@ -1,14 +1,14 @@
+[workspace]
+members = ["crates/format"]
+
 [package]
 name = "pactfmt"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+format = { path = "crates/format" }
 clap = { version = "4.5.20", features = ["derive"] }
-logos = "0.14.2"
-pretty = "0.12.3"
-test_each_file = "0.3.4"
-winnow = "0.6.20"
 
 [dev-dependencies]
 insta = "1.41.1"
