--- build.rs
@@ -1,3 +1,0 @@
-fn main() {
-    println!("cargo:rerun-if-changed=fixtures");
-}
