+++ crates/format/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "format"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+logos = "0.14.2"
+pretty = "0.12.3"
+winnow = "0.6.20"
+
+[dev-dependencies]
+insta = "1.41.1"
+test_each_file = "0.3.4"
