+++ crates/format/src/format.rs
@@ -642,35 +642,3 @@ where
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::cst;
-    use crate::format::FormatDoc;
-    use crate::lexer;
-    use crate::parser;
-
-    use logos::Logos;
-    use pretty::RcAllocator;
-
-    use test_each_file::test_each_file;
-
-    fn lex(input: &str) -> Vec<lexer::Token> {
-        lexer::Token::lexer(input)
-            .filter_map(|token| token.ok())
-            .collect()
-    }
-
-    test_each_file! { in "./fixtures" => |content: &str| {
-        let tokens = lex(content);
-        let mut input = tokens.as_slice();
-        let parsed = parser::parse(&mut input).expect("parsing");
-        let fst = parsed.into_iter().map(cst::lower_toplevel).collect::<Vec<_>>();
-
-        let doc_wide = fst.iter().fold(FormatDoc::nil(&RcAllocator), |acc, fst| acc.append(fst.format(&RcAllocator)));
-        let doc_narrow = fst.iter().fold(FormatDoc::nil(&RcAllocator), |acc, fst| acc.append(fst.format(&RcAllocator)));
-
-        insta::assert_snapshot!(doc_wide.pretty(80));
-        insta::assert_snapshot!(doc_narrow.pretty(20));
-    }}
-}
