+++ crates/format/src/lexer.rs
