+++ crates/format/tests/fixtures/defun_singleline.pact
